Re-source all scripts when invoking subshells
This serves two purposes:
* The tool bashcov correctly keeps track of coverage
* Stack traces show lines in source files

Several alternatives tried, the least dangerous seems to exhibit a bug in bash.

+++ test/test_FAIL_FAST.sh
@@ -1,5 +1,3 @@
-source "$(dirname "$(readlink -f "$0")")"/../test.sh
-
 test_ok() {
   touch "$TEST_SCRIPT_DIR"/.test_ok
   TEST_OK=1
@@ -12,6 +10,9 @@ test_fail() {
   false
   true
 }
+
+[ "$REENTRANT" != 1 ] || return 0
+source "$(dirname "$(readlink -f "$0")")"/../test.sh
 
 set_test_name "Any command that fails in the body of a test function should make the test to fail"
 rm -f "$TEST_SCRIPT_DIR"/.test_ok "$TEST_SCRIPT_DIR"/.test_fail
