+++ test/test_test_functions.sh
@@ -1,5 +1,3 @@
-source "$(dirname "$(readlink -f "$0")")"/../test.sh
-
 setup_test_suite() {
   echo setup_test_suite >>"$OUTFILE"
 }
@@ -26,6 +24,9 @@ test_02() {
   echo test_02 >>"$OUTFILE"
 }
 
+[ "$REENTRANT" != 1 ] || return 0
+source "$(dirname "$(readlink -f "$0")")"/../test.sh
+
 set_test_name "run_tests shoud invoke tests and setup methods when there are no failures"
 OUTFILE="$TEST_SCRIPT_DIR"/.test_test_functions.out
 rm -rf "$OUTFILE"
@@ -45,7 +46,7 @@ EOF
 set_test_name "run_tests shoud invoke tests and setup methods when there are failures"
 rm -rf "$OUTFILE"
 test_02_fail=1
-! CURRENT_TEST_NAME= bash -c run_tests
+! CURRENT_TEST_NAME= subshell run_tests
 
 diff - "$OUTFILE" <<EOF
 setup_test_suite
